extract new Notice module
