+++ lib/plausible_web/templates/auth/user_settings.html.heex
@@ -26,7 +26,7 @@
 
       <div class="my-4 border-b border-gray-400"></div>
 
-      <PlausibleWeb.Components.Billing.subscription_cancelled_notice
+      <PlausibleWeb.Components.Billing.Notice.subscription_cancelled
         user={@user}
         dismissable={false}
       />
@@ -311,7 +311,7 @@
     <h2 class="text-xl font-black dark:text-gray-100">API Keys</h2>
     <div class="my-4 border-b border-gray-300 dark:border-gray-500"></div>
 
-    <PlausibleWeb.Components.Billing.premium_feature_notice
+    <PlausibleWeb.Components.Billing.Notice.premium_feature
       billable_user={@current_user}
       current_user={@current_user}
       feature_mod={Plausible.Billing.Feature.StatsAPI}
