+++ lib/plausible_web/templates/site/membership/invite_member_form.html.heex
@@ -1,7 +1,7 @@
 <%= form_for @conn, Routes.membership_path(@conn, :invite_member, @site.domain), [class: "max-w-lg w-full mx-auto bg-white dark:bg-gray-800 shadow-md rounded px-8 pt-6 pb-8 mb-4 mt-8"], fn f -> %>
   <h2 class="text-xl font-black dark:text-gray-100 mb-4">Invite member to <%= @site.domain %></h2>
 
-  <PlausibleWeb.Components.Billing.limit_exceeded_notice
+  <PlausibleWeb.Components.Billing.Notice.limit_exceeded
     :if={Map.get(assigns, :is_at_limit, false)}
     current_user={@current_user}
     billable_user={@site.owner}
