+++ lib/plausible_web/templates/site/new.html.heex
@@ -2,7 +2,7 @@
   <%= form_for @changeset, "/sites", [class: "max-w-lg w-full mx-auto bg-white dark:bg-gray-800 shadow-lg rounded px-8 pt-6 pb-8 mb-4 mt-8"], fn f -> %>
     <h2 class="text-xl font-black dark:text-gray-100 mb-4">Your website details</h2>
 
-    <PlausibleWeb.Components.Billing.limit_exceeded_notice
+    <PlausibleWeb.Components.Billing.Notice.limit_exceeded
       :if={@site_limit_exceeded?}
       current_user={@current_user}
       billable_user={@current_user}
