+++ lib/plausible_web/templates/site/settings_funnels.html.heex
@@ -1,5 +1,5 @@
 <section class="shadow bg-white dark:bg-gray-800 sm:rounded-md sm:overflow-hidden">
-  <PlausibleWeb.Components.Billing.premium_feature_notice
+  <PlausibleWeb.Components.Billing.Notice.premium_feature
     billable_user={@site.owner}
     current_user={@current_user}
     feature_mod={Plausible.Billing.Feature.Funnels}
