+++ test/plausible_web/components/billing/billing_test.exs
@@ -1,23 +1,23 @@
 defmodule PlausibleWeb.Components.BillingTest do
   use Plausible.DataCase
   import Phoenix.LiveViewTest
-  alias PlausibleWeb.Components.Billing
+  alias PlausibleWeb.Components.Billing.Notice
 
-  test "premium_feature_notice/1 does not render a notice when user is on trial" do
+  test "premium_feature/1 does not render a notice when user is on trial" do
     me = insert(:user)
 
-    assert render_component(&Billing.premium_feature_notice/1,
+    assert render_component(&Notice.premium_feature/1,
              billable_user: me,
              current_user: me,
              feature_mod: Plausible.Billing.Feature.Props
            ) == ""
   end
 
-  test "premium_feature_notice/1 renders an upgrade link when user is the site owner and does not have access to the feature" do
+  test "premium_feature/1 renders an upgrade link when user is the site owner and does not have access to the feature" do
     me = insert(:user, subscription: build(:growth_subscription))
 
     rendered =
-      render_component(&Billing.premium_feature_notice/1,
+      render_component(&Notice.premium_feature/1,
         billable_user: me,
         current_user: me,
         feature_mod: Plausible.Billing.Feature.Props
@@ -28,12 +28,12 @@ defmodule PlausibleWeb.Components.BillingTest do
     assert rendered =~ "/billing/choose-plan"
   end
 
-  test "premium_feature_notice/1 does not render an upgrade link when user is not the site owner" do
+  test "premium_feature/1 does not render an upgrade link when user is not the site owner" do
     me = insert(:user)
     owner = insert(:user, subscription: build(:growth_subscription))
 
     rendered =
-      render_component(&Billing.premium_feature_notice/1,
+      render_component(&Notice.premium_feature/1,
         billable_user: owner,
         current_user: me,
         feature_mod: Plausible.Billing.Feature.Funnels
@@ -43,11 +43,11 @@ defmodule PlausibleWeb.Components.BillingTest do
     assert rendered =~ "please reach out to the site owner to upgrade their subscription"
   end
 
-  test "premium_feature_notice/1 does not render a notice when the user has access to the feature" do
+  test "premium_feature/1 does not render a notice when the user has access to the feature" do
     me = insert(:user, subscription: build(:business_subscription))
 
     rendered =
-      render_component(&Billing.premium_feature_notice/1,
+      render_component(&Notice.premium_feature/1,
         billable_user: me,
         current_user: me,
         feature_mod: Plausible.Billing.Feature.Funnels
@@ -56,11 +56,11 @@ defmodule PlausibleWeb.Components.BillingTest do
     assert rendered == ""
   end
 
-  test "limit_exceeded_notice/1 when billable user is on growth displays upgrade link" do
+  test "limit_exceeded/1 when billable user is on growth displays upgrade link" do
     me = insert(:user, subscription: build(:growth_subscription))
 
     rendered =
-      render_component(&Billing.limit_exceeded_notice/1,
+      render_component(&Notice.limit_exceeded/1,
         billable_user: me,
         current_user: me,
         limit: 10,
@@ -72,11 +72,11 @@ defmodule PlausibleWeb.Components.BillingTest do
     assert rendered =~ "/billing/choose-plan"
   end
 
-  test "limit_exceeded_notice/1 when billable user is on growth but is not current user does not display upgrade link" do
+  test "limit_exceeded/1 when billable user is on growth but is not current user does not display upgrade link" do
     me = insert(:user, subscription: build(:growth_subscription))
 
     rendered =
-      render_component(&Billing.limit_exceeded_notice/1,
+      render_component(&Notice.limit_exceeded/1,
         billable_user: me,
         current_user: insert(:user),
         limit: 10,
@@ -88,11 +88,11 @@ defmodule PlausibleWeb.Components.BillingTest do
   end
 
   @tag :full_build_only
-  test "limit_exceeded_notice/1 when billable user is on trial displays upgrade link" do
+  test "limit_exceeded/1 when billable user is on trial displays upgrade link" do
     me = insert(:user)
 
     rendered =
-      render_component(&Billing.limit_exceeded_notice/1,
+      render_component(&Notice.limit_exceeded/1,
         billable_user: me,
         current_user: me,
         limit: 10,
@@ -104,7 +104,7 @@ defmodule PlausibleWeb.Components.BillingTest do
     assert rendered =~ "/billing/choose-plan"
   end
 
-  test "limit_exceeded_notice/1 when billable user is on an enterprise plan displays support email" do
+  test "limit_exceeded/1 when billable user is on an enterprise plan displays support email" do
     me =
       insert(:user,
         enterprise_plan: build(:enterprise_plan, paddle_plan_id: "123321"),
@@ -112,7 +112,7 @@ defmodule PlausibleWeb.Components.BillingTest do
       )
 
     rendered =
-      render_component(&Billing.limit_exceeded_notice/1,
+      render_component(&Notice.limit_exceeded/1,
         billable_user: me,
         current_user: me,
         limit: 10,
@@ -123,11 +123,11 @@ defmodule PlausibleWeb.Components.BillingTest do
     assert rendered =~ "please contact hello@plausible.io to upgrade your subscription"
   end
 
-  test "limit_exceeded_notice/1 when billable user is on a business plan displays support email" do
+  test "limit_exceeded/1 when billable user is on a business plan displays support email" do
     me = insert(:user, subscription: build(:business_subscription))
 
     rendered =
-      render_component(&Billing.limit_exceeded_notice/1,
+      render_component(&Notice.limit_exceeded/1,
         billable_user: me,
         current_user: me,
         limit: 10,
